+++ Changelog.md
@@ -24,6 +24,7 @@ is replaced by the new portrait. (For unlinked Actors, this will directly affect
 -- **Stymied** - all actors on that faction's side are immediately given the Stymied (stacking) effect.
 - Drawing a Drama Card will auto-sort the combatants based on the new card and the Standard/Dramatic setting.
 - Drama Deck can be shuffled from the Combat Tracker's context menu (three dots to the right of the Round counter).
+- Combat Context Menu option to recall previous Drama Card.
 
 ### Bug Fixes